Extract record utils to a separate file
