+++ app/records/models.py
@@ -2,109 +2,26 @@ from __future__ import annotations
 
 import logging
 import re
-from abc import ABC, abstractmethod
-from typing import Any, Dict, Optional
+from typing import Any
 
+from app.records.utils import extract, format_link
 from django.urls import NoReverseMatch, reverse
 from django.utils.functional import cached_property
-from pyquery import PyQuery as pq
 
 from .converters import IDConverter
 
 logger = logging.getLogger(__name__)
 
 
-def format_link(link_html: str, inc_msg: str = "") -> Dict[str, str]:
-    """
-    Extracts iaid and text from a link HTML string, e.g. "<a href="C5789">DEFE 31</a>"
-    and returns as dict in the format: `{"id":"C5789", "href": "/catalogue/id/C5789/", "text":"DEFE 31"}
-
-    inc_msg includes message with logger if sepcified
-    Ex:inc_msg <method_name>:Record(<id):"
-    """
-    document = pq(link_html)
-    iaid = document.attr("href")
-    try:
-        href = reverse("details-page-machine-readable", kwargs={"iaid": iaid})
-    except NoReverseMatch:
-        href = ""
-        # warning for partially valid data
-        logger.warning(
-            f"{inc_msg}format_link:No reverse match for details-page-machine-readable with iaid={iaid}"
-        )
-    return {"id": iaid or "", "href": href, "text": document.text()}
-
-
-def extract(
-    source: Dict[str, Any], key: str, default: Optional[Any] = None
-) -> Any:
-    """
-    Attempts to extract `key` (a string with multiple '.' to indicate
-    traversal) from `source` (a complex multi-level dict where values may
-    by lists or other complex types) and return the value.
-
-    If `default` is provided, that value will be returned if any issues
-    arise during the process. When no `default` is provied, a
-    `ValueExtractionError` is raised instead.
-    """
-    current = source
-    lookups = tuple(key.split("."))
-
-    try:
-        for bit in lookups:
-            # NOTE: we could use a series of nested try/excepts here instead,
-            # but using conditionals allows us to raise more relevant exceptions
-
-            # Only attempt key lookups for dicts
-            if isinstance(current, dict):
-                current = current[bit]
-                continue
-
-            # Only attempt index lookups for sequences, and only
-            # when the value looks like an index
-            if hasattr(current, "__getitem__"):
-                try:
-                    bit_index = int(bit)
-                except ValueError:
-                    pass
-                else:
-                    current = current[bit_index]  # do index lookup
-                    continue
-
-            # Always fall back to attribute lookup
-            current = getattr(current, bit)
-
-    except Exception:
-        return default
-
-    return current
-
-
-class ValueExtractionError(Exception):
-    pass
-
-
 class APIResponse:
     def __init__(self, raw_data: dict[str, Any]):
         self._raw = raw_data
 
-    def get(self, key: str, default: Optional[Any] = None) -> Any:
-        """
-        Attempts to extract `key` from `self._raw` and return the value.
-        """
-        if "." in key:
-            try:
-                return extract(self._raw, key, default)
-            except ValueExtractionError:
-                return default
-        try:
-            return self._raw[key]
-        except KeyError:
-            return default
-
     @cached_property
-    def record(self):
-        return Record(self.get("@template.details"))
+    def record(self) -> Record | None:
+        if "@template" in self._raw and "details" in self._raw["@template"]:
+            return Record(self._raw["@template"]["details"])
+        return None
 
 
 class Record:
@@ -114,14 +31,14 @@ class Record:
     def __str__(self):
         return f"{self.summary_title} ({self.iaid})"
 
-    def get(self, key: str, default: Optional[Any] = None) -> Any:
+    def get(self, key: str, default: Any) -> Any:
         """
         Attempts to extract `key` from `self._raw` and return the value.
         """
         if "." in key:
             try:
                 return extract(self._raw, key, default)
-            except ValueExtractionError:
+            except Exception:
                 return default
         try:
             return self._raw[key]
@@ -190,7 +107,7 @@ class Record:
     @cached_property
     def summary_title(self) -> str:
         """Returns the api value of the attr if found, empty str otherwise."""
-        if details_summary_title := self.get("summaryTitle"):
+        if details_summary_title := self.get("summaryTitle", ""):
             return details_summary_title
         return self.get("summary.title", "")
 
@@ -237,8 +154,6 @@ class Record:
     @cached_property
     def level_code(self) -> int | None:
         """Returns the api value of the attr if found, None otherwise."""
-        if details_level_code := self.get("level.code"):
-            return details_level_code
         return self.get("level.code", "")
 
     @cached_property
